add unit tests, organise classes
