+++ src/main/java/uk/gov/companieshouse/api/accounts/model/entity/notes/stocks/StocksEntity.java
@@ -1,9 +1,7 @@
 package uk.gov.companieshouse.api.accounts.model.entity.notes.stocks;
 
-import org.springframework.data.mongodb.core.mapping.Document;
 import uk.gov.companieshouse.api.accounts.model.entity.NoteEntity;
 
-@Document(collection = "notes")
 public class StocksEntity extends NoteEntity {
 
     private StocksDataEntity data;
