+++ src/main/java/uk/gov/companieshouse/api/accounts/model/rest/smallfull/notes/stocks/PreviousPeriod.java
@@ -1,4 +1,4 @@
-package uk.gov.companieshouse.api.accounts.model.rest.notes.stocks;
+package uk.gov.companieshouse.api.accounts.model.rest.smallfull.notes.stocks;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
