+++ src/main/java/uk/gov/companieshouse/api/accounts/transformer/StocksTransformer.java
@@ -7,8 +7,8 @@ import uk.gov.companieshouse.api.accounts.model.entity.notes.stocks.CurrentPerio
 import uk.gov.companieshouse.api.accounts.model.entity.notes.stocks.PreviousPeriodEntity;
 import uk.gov.companieshouse.api.accounts.model.entity.notes.stocks.StocksDataEntity;
 import uk.gov.companieshouse.api.accounts.model.entity.notes.stocks.StocksEntity;
-import uk.gov.companieshouse.api.accounts.model.rest.notes.stocks.CurrentPeriod;
-import uk.gov.companieshouse.api.accounts.model.rest.notes.stocks.PreviousPeriod;
+import uk.gov.companieshouse.api.accounts.model.rest.smallfull.notes.stocks.CurrentPeriod;
+import uk.gov.companieshouse.api.accounts.model.rest.smallfull.notes.stocks.PreviousPeriod;
 import uk.gov.companieshouse.api.accounts.model.rest.smallfull.notes.stocks.Stocks;
 
 @Component
