+++ src/test/java/uk/gov/companieshouse/api/accounts/transformer/StocksTransformerTest.java
@@ -5,12 +5,13 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.TestInstance;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.junit.jupiter.MockitoExtension;
+import uk.gov.companieshouse.api.accounts.enumeration.AccountingNoteType;
 import uk.gov.companieshouse.api.accounts.model.entity.notes.stocks.CurrentPeriodEntity;
 import uk.gov.companieshouse.api.accounts.model.entity.notes.stocks.PreviousPeriodEntity;
 import uk.gov.companieshouse.api.accounts.model.entity.notes.stocks.StocksDataEntity;
 import uk.gov.companieshouse.api.accounts.model.entity.notes.stocks.StocksEntity;
-import uk.gov.companieshouse.api.accounts.model.rest.notes.stocks.CurrentPeriod;
-import uk.gov.companieshouse.api.accounts.model.rest.notes.stocks.PreviousPeriod;
+import uk.gov.companieshouse.api.accounts.model.rest.smallfull.notes.stocks.CurrentPeriod;
+import uk.gov.companieshouse.api.accounts.model.rest.smallfull.notes.stocks.PreviousPeriod;
 import uk.gov.companieshouse.api.accounts.model.rest.smallfull.notes.stocks.Stocks;
 
 import java.util.HashMap;
@@ -143,6 +144,16 @@ public class StocksTransformerTest {
         assertEquals(ETAG, stocks.getEtag());
         assertEquals(KIND, stocks.getKind());
     }
+
+    @Test
+    @DisplayName("transformer returns stocks note")
+    void testStockNoteReturned() {
+
+        AccountingNoteType noteType = AccountingNoteType.SMALL_FULL_STOCKS;
+
+        assertEquals(noteType, stocksTransformer.getAccountingNoteType());
+    }
+
 
     private PreviousPeriodEntity createPreviousPeriodEntityObject() {
